+++ ruby/CHANGELOG.md
@@ -8,9 +8,11 @@ This is the changelog for the `essence` gem. For now, most of the updates will b
 - feature: Adding `Card` component
 - feature: Adding `Checkbox` component
 - feature: Adding `Input` component
+- feature: Adding `Heading` component
 - feature: Adding `Label` component
 - feature: Adding `Separator` component
 - feature: Adding `Switch` component
+- feature: Adding `Text` component
 - fix: Removing `Phlex::Kit` from the base component
 - fix: Removing unecessary autoloading